Clean up generated SVG
Two changes: (1) Apply our default style to the group as a whole, not
to the individual path elements. This results in smaller generated SVG,
especially where the extension is used heavily.  (2) If the text
"rendered" consists only of whitespace -- that is, if no actual paths
were rendered -- then no output (no empty group) is added to the SVG
document.