Adds some extra options to eggbot_hatch. These are especially usefull when working with generated content that has a lot of elements you don't want to have to alter one-by-one.
First is the option to delete the original object that was hatched.
This allows you to only plot the newly made hatched fill, not the
outline of the original object.

Second is an option to use ranges for the hatchAngle and hatchSpacing
variables. For each element it's brightness is calculated (0.1-1.0,
where 0.0 is a black element and 1.0 is a white one) and this brightness
is then mapped to the range. So if you select 0 to 30 for your angle
range, black elements will use 0 degrees, white elements will use 30
and gray elements can use any value in between. This allows for
the creation of some nice effects.